Added check if rigging needs updating...
I found that switching to an editing stage with a lot of rigging on it was causing Blender to hang for a couple of seconds while it updated... i've added a "does this rigging need to be updated" check to allievate some of the strain and only have rigging update when it needs to. It was mostly the spline and plantigrade chains causing the stress as their code is quite heavy to make them agnostic, but this should generally improve performance anyways.