Much improved modular code...
WARNING: There may still be a few bugs and the "default shapes" option will not work until i do a release .zip that includes the library .blend for them.

I re-wrote the entire rigging library add-on to make it faster and less fragile and also so i can more easily add new bits of rigging to it.